Escrow load service + settlement table update _ readme.md updated

+++ src/main/java/org/payment/util/CurrencyConverter.java
@@ -6,7 +6,7 @@ public interface CurrencyConverter {
 
     BigInteger toMinorUnit(float majorAmount);      // ₹10.50 → 1050
 
-    float toMajorUnit(BigInteger minorAmount);     // 1050 → ₹10.50
+    float toMajorUnit(long minorAmount);     // 1050 → ₹10.50
 
     String getCurrencyCode();
 }
